Add runtime support for Linux/ARM/Raspberry Pi
The runtime executable can now be cross-compiled for Raspberry Pi. None of the other targets (compiler or console) currently support cross-compilation. This change includes the following sub-changes:
- Fix a number of warnings/errors caught by gcc
- Jamfiles added for Raspberry Pi
- Add NEON implementation of the SIMD math classes and functions
- A more expressive assert message is now displayed
- Fixed latency bug in audio driver interface
- Fixed bug in async event handler
- Fixed bug where buffers were being allocated 8x too large
- Thread pool now runs tasks synchronously on the controlling thread when a worker count of 0 is specified
- A default runtime config file with inline documentation is generated if none is found on startup
- Fixed minor bug in profiler when no voices are ever activated
- Fixed bug where #activate_fx_immediately was not ever stored
- Added math unit tests

+++ source/engine/task_functions/task_functions_delay.cpp
@@ -108,7 +108,7 @@ size_t s_buffer_operation_delay::query_memory(uint32 sample_rate, real32 delay) 
 	}
 
 	uint32 delay_samples = static_cast<uint32>(
-		std::max(0.0f, round(static_cast<real32>(sample_rate) * delay)));
+		std::max(0.0f, roundf(static_cast<real32>(sample_rate) * delay)));
 
 	// Allocate the size of the context plus the delay buffer
 	return sizeof(s_buffer_operation_delay) + (delay_samples * sizeof(real32));
@@ -122,7 +122,7 @@ void s_buffer_operation_delay::initialize(c_event_interface *event_interface,
 	}
 
 	context->delay_samples = static_cast<uint32>(
-		std::max(0.0f, round(static_cast<real32>(sample_rate) * delay)));
+		std::max(0.0f, roundf(static_cast<real32>(sample_rate) * delay)));
 	context->delay_buffer_head_index = 0;
 	context->is_constant = true;
